+++ include/YODA/Profile1D.h
@@ -114,7 +114,7 @@ namespace YODA {
 
 
     /// Fill dimension of this data object
-    int dim() const { return 1; }
+    size_t dim() const { return 1; }
 
 
     /// @name Modifiers
