+++ include/YODA/Scatter3D.h
@@ -138,7 +138,7 @@ namespace YODA {
 
 
     /// Dimension of this data object
-    int dim() const { return 3; }
+    size_t dim() const { return 3; }
 
 
     /// @name Modifiers
